클립 시간을 time(3) 으로 — 초 절삭이 샷 경계를 넘기던 문제
t_compose_clip.start_time/end_time 이 TIME(정수 초)이라 4369.3s 가 4369s 로
내려가 직전 샷에 걸쳤다. "투구에서 시작한 클립 0%" 로 보이던 게 실은 이 착시였고,
실제로는 75~92% 가 투구 시작이었다.

컬럼을 time(3) 으로 넓히고(ALTER 실행 완료, 기존 행 값 불변), fetch 는
TIME_TO_SEC 의 Decimal 을 float 으로 통일한다 — int() 로 내리면 원인이 돌아온다.

파급: ui-sbs-viwer 의 TIME_TO_SEC(cc.start_time) 이 DECIMAL 을 받게 된다
(별도 레포, 미반영).

+++ src/db/compose_repo.py
@@ -103,5 +103,8 @@ class ComposeRepo:
             clips = list(await cur.fetchall())
         head["reg_datetime"] = str(head.get("reg_datetime"))
         for c in clips:
-            c["start"], c["end"] = int(c["start"]), int(c["end"])
+            # TIME_TO_SEC 은 Decimal 을 준다 — float 로 통일한다(파이프라인 좌표=초).
+            # int() 로 내리면 안 된다: 컬럼이 time(3) 이라 밀리초가 살아 있고, 그
+            # 0.3초가 샷 경계를 가른다 (bounds._match 참조).
+            c["start"], c["end"] = float(c["start"]), float(c["end"])
         return {**head, "clips": clips}
